# docs/swagger.yml
openapi: 3.0.3
info:
  version: 1.0.0
  title: API Laboratorie Xams
  description: API para manutenção de laboratórios e exames
  contact:
    name: API Support
    email: murilosantoseduardo@gmail.com
servers:
  - description: Desenvolvimento
    url: http://localhost:3333
...
paths:
...
  /associations/{laboratory_id}:
    post:
      tags:
        - Association
      description: Associar um exame ativo à um laboratório ativo
      summary: Associação de exames à laboratórios
      parameters:
        - in: path
          name: laboratory_id
          schema:
            type: integer
          required: true
          description: Identificador do laboratório a ser vinculado
      responses:
        204:
          description: Associação efetuada com sucesso
        500:
          description: Erro interno do servidor
    delete:
      tags:
        - Association
      description: Desassociar um exame ativo à um laboratório ativo
      summary: Desassociação de exames à laboratórios
      parameters:
        - in: path
          name: laboratory_id
          schema:
            type: integer
          required: true
          description: Identificador do laboratório a ser desvinculado
      responses:
        204:
          description: Desassociação efetuada com sucesso
        500:
          description: Erro interno do servidor
...